+++ src/irvine/oeis/a023/A023871.java
@@ -12,6 +12,6 @@ public class A023871 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A023871() {
-    super(0, new Builder().f(k -> Z.valueOf(k * k)));
+    super(0, new Builder().f(k -> Z.valueOf((long) k * k)));
   }
 }
