+++ src/irvine/oeis/a028/A028377.java
@@ -12,6 +12,6 @@ public class A028377 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A028377() {
-    super(0, new Builder().f(k -> Z.valueOf(-(k * (k + 1) / 2))).g(-1));
+    super(0, new Builder().f(k -> Z.valueOf(-((long) k * (k + 1) / 2))).g(-1));
   }
 }
