+++ src/irvine/oeis/a210/A210363.java
@@ -14,6 +14,6 @@ public class A210363 extends FilterSequence {
 
   /** Construct the sequence. */
   public A210363() {
-    super(1, new A000040(), p -> !Predicates.PRIME.is(p.add(5 * (5 + 1))) && Integers.SINGLETON.count(0, 4, x -> !Predicates.PRIME.is(p.add(x * (x + 1)))).isZero());
+    super(1, new A000040(), p -> !Predicates.PRIME.is(p.add(5 * (5 + 1))) && Integers.SINGLETON.count(0, 4, x -> !Predicates.PRIME.is(p.add((long) x * (x + 1)))).isZero());
   }
 }
