+++ src/irvine/oeis/a210/A210364.java
@@ -14,6 +14,6 @@ public class A210364 extends FilterSequence {
 
   /** Construct the sequence. */
   public A210364() {
-    super(1, new A000040(), p -> !Predicates.PRIME.is(p.add(6 * (6 + 1))) && Integers.SINGLETON.count(0, 5, x -> !Predicates.PRIME.is(p.add(x * (x + 1)))).isZero());
+    super(1, new A000040(), p -> !Predicates.PRIME.is(p.add(6 * (6 + 1))) && Integers.SINGLETON.count(0, 5, x -> !Predicates.PRIME.is(p.add((long) x * (x + 1)))).isZero());
   }
 }
