+++ src/irvine/oeis/a258/A258351.java
@@ -12,6 +12,6 @@ public class A258351 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A258351() {
-    super(0, new Builder().f(k -> Z.valueOf(k * (k - 1) * (k - 2))));
+    super(0, new Builder().f(k -> Z.valueOf((long) k * (k - 1) * (k - 2))));
   }
 }
