+++ src/irvine/oeis/a279/A279360.java
@@ -12,6 +12,6 @@ public class A279360 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A279360() {
-    super(0, new Builder().f(-1).gl(k -> -2L).h(k -> Z.valueOf(k * k)));
+    super(0, new Builder().f(-1).gl(k -> -2L).h(k -> Z.valueOf((long) k * k)));
   }
 }
