+++ src/irvine/oeis/a280/A280129.java
@@ -12,6 +12,6 @@ public class A280129 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A280129() {
-    super(0, new Builder().f(-1).g(k -> (k < 2) ? Z.ZERO : Z.NEG_ONE).h(k -> Z.valueOf(k * k)));
+    super(0, new Builder().f(-1).g(k -> (k < 2) ? Z.ZERO : Z.NEG_ONE).h(k -> Z.valueOf((long) k * k)));
   }
 }
