+++ src/irvine/oeis/a285/A285674.java
@@ -12,6 +12,6 @@ public class A285674 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A285674() {
-    super(0, new Builder().fl(k -> k).g(k -> Z.valueOf(k * k)));
+    super(0, new Builder().fl(k -> k).g(k -> Z.valueOf((long) k * k)));
   }
 }
