+++ src/irvine/oeis/a292/A292892.java
@@ -14,6 +14,6 @@ public class A292892 extends LambdaArray {
 
   /** Construct the sequence. */
   public A292892() {
-    super(0, 0, 0, -1, (n, k) -> Rationals.SINGLETON.sum(0, n / (k + 1), j -> new Q(Functions.STIRLING2.z(n - k * j, j), Functions.FACTORIAL.z(n - k * j))).multiply(Functions.FACTORIAL.z(n)).num());
+    super(0, 0, 0, -1, (n, k) -> Rationals.SINGLETON.sum(0, n / (k + 1), j -> new Q(Functions.STIRLING2.z(n - (long) k * j, j), Functions.FACTORIAL.z(n - k * j))).multiply(Functions.FACTORIAL.z(n)).num());
   }
 }
