+++ src/irvine/oeis/a294/A294581.java
@@ -12,6 +12,6 @@ public class A294581 extends RationalProductTransform {
 
   /** Construct the sequence. */
   public A294581() {
-    super(0, new Builder().fl(k -> -k).g(k -> Z.valueOf(k * k)));
+    super(0, new Builder().fl(k -> -k).g(k -> Z.valueOf((long) k * k)));
   }
 }
